modularize abs, acos, acosh, asin, asinh, atan, atanh
