+++ tfjs-core/src/tests.ts
@@ -38,7 +38,10 @@ import './io/router_registry_test';
 import './io/weights_loader_test';
 import './jasmine_util_test';
 import './kernel_registry_test';
+import './ops/abs_test';
 import './ops/absolute_difference_test';
+import './ops/acos_test';
+import './ops/acosh_test';
 import './ops/add_n_test';
 import './ops/add_test';
 import './ops/all_test';
@@ -47,6 +50,10 @@ import './ops/arg_max_test';
 import './ops/arg_min_test';
 import './ops/arithmetic_test';
 import './ops/array_ops_test';
+import './ops/asin_test';
+import './ops/asinh_test';
+import './ops/atan_test';
+import './ops/atanh_test';
 import './ops/avg_pool_3d_test';
 import './ops/avg_pool_test';
 import './ops/axis_util_test';
@@ -163,6 +170,7 @@ import './ops/topk_test';
 import './ops/transpose_test';
 import './ops/truncated_normal_test';
 import './ops/unary_ops_test';
+import './ops/undefined_test';
 import './ops/unsorted_segment_sum_test';
 import './ops/unstack_test';
 import './ops/where_async_test';
